+++ src/main.c
@@ -1,64 +1,38 @@
 /**
  * @file main.c
- * @brief Application entry point demonstrating driven GPIO handling.
+ * @brief Application entry point demonstrating driver interaction.
  *
- * Configures:
- * - GPIO for LED output and button input
  *
  * Behavior:
- * - Button press while polling
- * - While pressed, set LED on, else, off
+ * - Polls a I2C device's ID.
+ * - Logs this ID to the console on PC via USART/ST-Link.
  *
  * Written by Edwin J Martinez
  */
 
-#include "driver.h"
-
+#include "hardware.h"
+#include "i2c_driver.h"
+#include "logger.h"
 #include <stdbool.h>
-
-//* Pins
-#define LED_PIN             0x00U
-
-Button g_Button = 
-{
-    port:           GPIOA_REGS,
-    pin:            0x01U,
-    isActiveLow:    true
-};
-
-
-void init()
-{
-    // Enable the clock for Port A
-    enable_gpio_clock(GPIOAEN);
-
-    // Set LED pin, MODER, to Output Mode
-    set_gpio_mode(GPIOA_REGS, LED_PIN, GPIOMODE_OUTPUT);
-
-    // Set Button Pin, MODER, to Input Mode
-    set_gpio_mode(GPIOA_REGS, g_Button.pin, GPIOMODE_INPUT);
-
-    // Set Button Pin, PUPDR, to Pull-Up Resistor
-    set_pupd_mode(GPIOA_REGS, g_Button.pin, PULL_UP);
-
-    // Set LED to off by default
-    clear_pin_by_bsrr(GPIOA_REGS, LED_PIN);
-
-}
 
 int main()
 {
-    init();
+
+    init_logging(115200);
+
+    i2c_init();
+
+    // Wake up MPU6050 (MPU6050's address is 0x68. Its reset register address is 0x6B and we just send empty data 0x00)
+    i2c_write_register(I2C1_REGS, 0x68, 0x6B, 0x00);
+
+    // Read identity register (MPU6050's address is 0x68 and its WHO_AM_I register is at 0x75)
+    uint8_t id = i2c_read_register(I2C1_REGS, 0x68, 0x75);
 
     while (1)
     {
-        // simple button press & hold to turn light on
-        bool isButtonPressed = is_button_pressed(&g_Button);
 
-        if (isButtonPressed)
-            set_pin_by_bsrr(GPIOA_REGS, LED_PIN);
-        else
-            clear_pin_by_bsrr(GPIOA_REGS, LED_PIN);  
+        for (volatile uint32_t i = 0; i < 500000; i++) {} // simple delay so printing is not spammed
+        prints("I2C MPU-6050 device id: %d\r\n", id);
     }
     
 }